// JeeFlatForm-MicroFrontend-/projects/jeeworkflow/src/app/page/models/jee-work.model.ts
export class ToDoDataModel extends BaseModel {
	RowID: number;
	TaskName: string;
	Description: string;
	NumberofTime: string;
	Time_Type: string;
	Data_Implementer: any[] = [];
	Data_Follower: any[] = [];
	Priority: number;
	NodeID: number;
	clear() {
		this.RowID = 0;
		this.TaskName = "";
		this.Description = ""
		this.NumberofTime = "";
		this.Time_Type = "";
		this.Data_Implementer = [];
		this.Data_Follower = [];
		this.Priority = 0;
		this.NodeID;
	}
}